增加了导入图片功能 增加GL资源的持久化功能
Former-commit-id: 4385d4f76c8bdeeb4767cf7739cebcc6048b3feb
Former-commit-id: 06059155cfb08680d7b3c5aaddbe2778780e01ba

+++ include/BDXKEngine/Base/Data/Mathematics/Matrix/Matrix4x4.cpp
@@ -28,7 +28,7 @@ namespace BDXKEngine
     }
     Matrix4x4 Matrix4x4::Rotate(Vector3 degree)
     {
-        const Vector3 radian = degree / 180 * M_PI;
+        const Vector3 radian = degree / static_cast<float>(180 * M_PI);
 
 
         const float ry = radian.y;
